requirements.txt
fastapi==0.87.0
uvicorn[standard]==0.20.0
pydantic==1.10.2
python-dotenv==0.21.0
SQLAlchemy==1.4.11
pre-commit==2.20.0
psycopg2-binary==2.9.5